refactor(web): extract components
- Header
- NavbarItems
