+++ web/components/Layout/NavbarItems/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./NavbarItems";
