--- web/components/MobileMenu/MobileMenu.module.scss
@@ -1,112 +1,0 @@
-@import "styles/vars.scss";
-@import "../Layout/Layout.module.scss";
-
-// mobile menu
-.menu {
-  background-color: $bg-dark;
-  position: fixed;
-  top: $mobile-header-height;
-  z-index: 10;
-  width: 100%;
-  height: calc(100% - $mobile-header-height);
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  //   touch-action: none;
-  overflow-y: scroll;
-  overflow-x: hidden;
-
-  nav {
-    margin-top: 32px;
-    margin-left: 20px;
-    line-height: 37.5px;
-    li {
-      &:hover {
-        cursor: pointer;
-      }
-    }
-    .category {
-      margin-top: calc(32px - 2px);
-      .category_item {
-        display: inline-flex;
-        justify-content: center;
-        align-items: center;
-        padding: 2px 9px;
-        padding-right: 0;
-        transition: background-color 200ms linear;
-      }
-
-      // rotate the arrow
-      img {
-        transition: transform 200ms linear;
-        transform: rotate(-90deg);
-      }
-
-      &.collapsed {
-        // rotate the arrow
-        img {
-          transform: rotate(0);
-        }
-        .category_item {
-          margin-bottom: 0.3rem;
-        }
-      }
-
-      &.active {
-        .category_item {
-          color: rgba(255, 255, 255, 1);
-          background-color: rgba(255, 255, 255, 0.07);
-          border-radius: 8px;
-        }
-      }
-      .menu_list {
-        position: relative;
-        left: 1.3rem;
-      }
-    }
-    .menu_list {
-      all: unset;
-      list-style-type: none;
-      font-family: "Roboto Condensed";
-      font-weight: 500;
-      font-size: 2rem;
-      color: rgba(255, 255, 255, 0.75);
-
-      li:not(:first-child) {
-        margin-top: 0.7rem;
-      }
-    }
-
-    .link {
-      padding: 2px 0;
-      padding-left: 9px;
-      a {
-        display: block;
-      }
-      &.active,
-      &:hover {
-        color: $primary-color;
-      }
-    }
-  }
-
-  .join {
-    all: unset;
-    &:hover {
-      cursor: pointer;
-    }
-    margin: 0 auto;
-    margin-bottom: 34px;
-    margin-top: 30px;
-    padding: 10px 55px;
-    font-family: "Roboto Condensed";
-    font-weight: 500;
-    font-size: 1.5rem;
-    filter: drop-shadow(0px 1px 4px #4ca7dd);
-    border-radius: 8px;
-    border: 1.5px solid #5fc3ff;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-}
